Phase 6d: carve demo mode out of app.js into demo.ts
demo.ts owns the guided-demo state machine: DEMO_STEPS (typed with a
DemoAction union for the interactive wait-for-* steps), the overlay/
hint rendering, positioning, dragging, and step navigation. Screen
switching goes through window.showScreen with screens resolved by id;
app.js passes its resetAndGoHome callback via initDemo() (which also
wires the Try Demo button, replacing the self-wiring IIFE).

The three app.js hooks now call demoActionCompleted('wait-for-start' /
'wait-for-submit') directly; the vote hook keeps its all-judges-voted
check locally (it needs app.js vote state) and gates on
currentDemoAction().

Verified: build + tsc + eslint clean; 9-step Playwright smoke passed;
targeted browser check of the demo flow (overlay + prefill, Next/
Previous, interactive wait-for-start auto-advance into the battle
screen, Exit Demo reset to home) — zero JS errors.
